==> 01_Hi-C/scripts/00_balanced_matrices.Rmd <==
---
title: "balanced_matrices"
author: "Elsie Jacobson"
date: "7/9/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(plyranges)


show_col(viridis_pal(option="C", begin=0.4, end=0.6)(20))

mypal <- c("#414487FF","#7AD151FF","#CC4977FF")
ggformat <- theme_classic(base_family = "Arial") + theme(text = element_text(size = 20))
```

```{r}
dist_cont <- read.csv("../data/ICE_distance_contact_prob.csv")
```

```{r}
ggplot(dist_cont, aes(distance, proportion, colour=condition)) + geom_line(size=0.5) + scale_y_log10() + scale_x_log10() + coord_cartesian(ylim=c(3e-04,0.15), xlim=c(100000,1e+08)) + scale_colour_manual(values=mypal, labels=c("Unmigrated","14µm pores","5µm pores")) + geom_rect(aes(xmin=100000,xmax=500000,ymin=0.03,ymax=0.15), fill=NA,colour="grey",linetype="dashed") + ggformat + labs(x="Distance", y="Proportion of contacts",colour="") + theme(legend.position=c(0.2,0.3))
ggsave("../plots/ICE_dist_contact_prob.pdf",width=8)

ggplot(dist_cont, aes(distance, proportion, colour=condition)) + geom_line() + scale_y_log10() + scale_x_log10(breaks=c(1e+05,2e+05,5e+05)) + coord_cartesian(ylim=c(0.03,0.15), xlim=c(100000,500000)) + scale_colour_manual(values=mypal, labels=c("Unmigrated","14µm pores","5µm pores")) + theme_classic(base_family = "Arial") + theme(text = element_text(size = 35.55556)) + labs(x="",y="", colour="")
ggsave("../plots/ICE_dist_contact_prob_zoom.pdf", width=10)
```

```{r}
contC <- dist_cont %>%
  filter(condition=="control")

cont5 <- dist_cont %>%
  filter(condition=="mig5")

cont14 <- dist_cont %>%
  filter(condition=="mig14")

ks.test(cont14$proportion, contC$proportion) ## not significant

ks.test(cont5$proportion, contC$proportion) ## significant!!

ks.test(cont5$proportion, cont14$proportion) ## significant!!
```
